cmd/classifier/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
ultralytics
gradio
fastapi
pillow
uvicorn
requests
pycryptodome